+++ native/packages/react-native-bpk-component-button-link/src/BpkButtonLink.ios.js
@@ -43,6 +43,7 @@ export type Props = {
 const BpkButtonLink = (props: Props) => {
   const {
     accessibilityLabel,
+    disabled,
     icon,
     iconAlignment,
     large,
@@ -63,6 +64,8 @@ const BpkButtonLink = (props: Props) => {
   const textStyle = [styles.text];
   const iconStyle = [styles.icon];
 
+  const accessibilityTraits = ['button'];
+
   if (iconAlignment === ICON_ALIGNMENTS.leading) {
     viewStyle.push(styles.viewLeading);
     iconStyle.push(styles.iconLeading);
@@ -71,6 +74,11 @@ const BpkButtonLink = (props: Props) => {
   if (themeStyle) {
     textStyle.push(themeStyle);
     iconStyle.push(themeStyle);
+  }
+
+  if (disabled) {
+    textStyle.push(styles.disabled);
+    accessibilityTraits.push('disabled');
   }
 
   if (style) {
@@ -82,8 +90,9 @@ const BpkButtonLink = (props: Props) => {
       <TouchableOpacity
         accessibilityComponentType="button"
         accessibilityLabel={accessibilityLabel || title}
-        accessibilityTraits={['button']}
+        accessibilityTraits={accessibilityTraits}
         onPress={onPress}
+        disabled={disabled}
         icon={icon}
         {...rest}
       >