+++ native/packages/react-native-bpk-component-button-link/stories.js
@@ -50,6 +50,13 @@ const createButtonStory = () => {
       />
       <BpkButtonLink
         large
+        disabled
+        title="Disabled"
+        onPress={action(`This should not be possible`)}
+        style={styles.buttonLinkStyles}
+      />
+      <BpkButtonLink
+        large
         title="With Leading icon"
         icon="alert--active"
         iconAlignment="leading"
@@ -73,6 +80,12 @@ const createButtonStory = () => {
       <BpkButtonLink
         title="Button"
         onPress={action(`Button pressed`)}
+        style={styles.buttonLinkStyles}
+      />
+      <BpkButtonLink
+        disabled
+        title="Disabled"
+        onPress={action('This should not be possible')}
         style={styles.buttonLinkStyles}
       />
       <BpkButtonLink